add basic pagination system

+++ view/frontend/moderationPage.php
@@ -18,12 +18,21 @@
     
     <div id='blog-comments' class="container">
         <?php
-            for($i = 0; $i < sizeof($comments); $i++){
-                echo "<div class='commentGroup'>";
-                    echo "<div class='comment node-depth-".$comments[$i]['depth']."' id='post".$comments[$i]['id']."'>";
-                    echo    "<p class='comment-content'>".$comments[$i]['comment']."</p>";
-                    echo    "<p class='comment-signature'>Rédigé par: <strong>".$comments[$i]['author']."</strong>, le [".$comments[$i]['comment_date']."] <button onclick='displayForm(".$comments[$i]['id'].",".$comments[$i]['depth'].")' class='comment-answer btn btn-primary' data-comment-id=".$comments[$i]['id'].">Répondre</button><button)' class='btn-report btn btn-warning'><i class='far fa-bell'></i> Signaler</button></p><br>";
-                echo "</div>";
+            if(isset($_GET['page'])){
+                $page = $_GET['page'];
+            }
+            else{
+
+                $page = 1;
+            }
+            for($i = 0+($sizePage*($page-1)); $i < $sizePage*$page; $i++){
+                if(array_key_exists($i, $comments)){
+                    echo "<div class='commentGroup'>";
+                        echo "<div class='comment node-depth-".$comments[$i]['depth']."' id='post".$comments[$i]['id']."'>";
+                        echo    "<p class='comment-content'>".$comments[$i]['comment']."</p>";
+                        echo    "<p class='comment-signature'>Rédigé par: <strong>".$comments[$i]['author']."</strong>, le [".$comments[$i]['comment_date']."] <button onclick='displayForm(".$comments[$i]['id'].",".$comments[$i]['depth'].")' class='comment-answer btn btn-primary' data-comment-id=".$comments[$i]['id'].">Répondre</button><button)' class='btn-report btn btn-warning'><i class='far fa-bell'></i> Signaler</button></p><br>";
+                    echo "</div>";
+                }
             }
         ?>
             
@@ -33,7 +42,7 @@
                 <li class="page-item"><a class="page-link" href="#">Previous</a></li>
                 <?php
                 for($i = 1; $i <= $nbPage ;$i++){
-                    echo "<li class='page-item'><a class='page-link' href='#'>".$i."</a></li>";
+                    echo "<li class='page-item'><a class='page-link' href='index.php?action=moderation&page=".$i."'>".$i."</a></li>";
                 }
                 ?>
                 <li class="page-item"><a class="page-link" href="#">Next</a></li>
